[kong] add workaround for leftover stream socket
Kong 2.3 added a unix socket for communication between the http and
stream systems. Per https://github.com/Kong/kong/issues/6827 this socket
file sometimes remains after Kong stops, which prevents the chart from
starting Kong when the wait-for-db init container runs (when using
Postgres or Cassandra).

This workaround deletes the problem file after wait-for-db stops Kong.
It should be removed once the bug is fixed, though it won't cause issues
if left in place (the rm will just be a no-op).